requirements.txt
streamlit
pdfplumber
rapidfuzz
scikit-learn
numpy
rake-nltk
sumy
nltk
spacy
textblob
wordcloud
matplotlib
pandas
pillow
